Refactor user preferences handling to utilize list page size from user settings
- Updated UserGeneralPreferences to change list_page_size from i32 to u64 for better compatibility with user preferences.
- Modified MiscellaneousService and CollectionService to incorporate user-specific list page size in pagination logic.
- Ensured compliance with coding standards by maintaining the order of attributes based on their length.

+++ crates/models/user/src/lib.rs
@@ -409,7 +409,7 @@ pub struct UserGeneralPreferences {
     #[educe(Default = "/")]
     pub landing_path: String,
     #[educe(Default = 20)]
-    pub list_page_size: i32,
+    pub list_page_size: u64,
     #[educe(Default = false)]
     pub disable_videos: bool,
     #[educe(Default = false)]
